+++ table/transaction.go
@@ -25,145 +25,10 @@ import (
 	"sync"
 	"time"
 
-	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/apache/iceberg-go"
 	"github.com/apache/iceberg-go/io"
-	"github.com/pterm/pterm"
 )
-
-type schemaCompatVisitor struct {
-	provided *iceberg.Schema
-
-	errorData pterm.TableData
-}
-
-func checkSchemaCompat(requested, provided *iceberg.Schema) error {
-	sc := &schemaCompatVisitor{
-		provided:  provided,
-		errorData: pterm.TableData{{"", "Table Field", "Requested Field"}},
-	}
-
-	_, compat := iceberg.PreOrderVisit(requested, sc)
-
-	return compat
-}
-
-func checkArrowSchemaCompat(requested *iceberg.Schema, provided *arrow.Schema, downcastNanoToMicro bool) error {
-	mapping := requested.NameMapping()
-	providedSchema, err := ArrowSchemaToIceberg(provided, downcastNanoToMicro, mapping)
-	if err != nil {
-		return err
-	}
-
-	return checkSchemaCompat(requested, providedSchema)
-}
-
-func (sc *schemaCompatVisitor) isFieldCompat(lhs iceberg.NestedField) bool {
-	rhs, ok := sc.provided.FindFieldByID(lhs.ID)
-	if !ok {
-		if lhs.Required {
-			sc.errorData = append(sc.errorData,
-				[]string{"❌", lhs.String(), "missing"})
-
-			return false
-		}
-		sc.errorData = append(sc.errorData,
-			[]string{"✅", lhs.String(), "missing"})
-
-		return true
-	}
-
-	if lhs.Required && !rhs.Required {
-		sc.errorData = append(sc.errorData,
-			[]string{"❌", lhs.String(), rhs.String()})
-
-		return false
-	}
-
-	if lhs.Type.Equals(rhs.Type) {
-		sc.errorData = append(sc.errorData,
-			[]string{"✅", lhs.String(), rhs.String()})
-
-		return true
-	}
-
-	// we only check that parent node is also of the same type
-	// we check the type of the child nodes as we traverse them later
-	switch lhs.Type.(type) {
-	case *iceberg.StructType:
-		if rhs, ok := rhs.Type.(*iceberg.StructType); ok {
-			sc.errorData = append(sc.errorData,
-				[]string{"✅", lhs.String(), rhs.String()})
-
-			return true
-		}
-	case *iceberg.ListType:
-		if rhs, ok := rhs.Type.(*iceberg.ListType); ok {
-			sc.errorData = append(sc.errorData,
-				[]string{"✅", lhs.String(), rhs.String()})
-
-			return true
-		}
-	case *iceberg.MapType:
-		if rhs, ok := rhs.Type.(*iceberg.MapType); ok {
-			sc.errorData = append(sc.errorData,
-				[]string{"✅", lhs.String(), rhs.String()})
-
-			return true
-		}
-	}
-
-	if _, err := iceberg.PromoteType(rhs.Type, lhs.Type); err != nil {
-		sc.errorData = append(sc.errorData,
-			[]string{"❌", lhs.String(), rhs.String()})
-
-		return false
-	}
-
-	sc.errorData = append(sc.errorData,
-		[]string{"✅", lhs.String(), rhs.String()})
-
-	return true
-}
-
-func (sc *schemaCompatVisitor) Schema(s *iceberg.Schema, v func() bool) bool {
-	if !v() {
-		pterm.DisableColor()
-		tbl := pterm.DefaultTable.WithHasHeader(true).WithData(sc.errorData)
-		tbl.Render()
-		txt, _ := tbl.Srender()
-		pterm.EnableColor()
-		panic("mismatch in fields:\n" + txt)
-	}
-
-	return true
-}
-
-func (sc *schemaCompatVisitor) Struct(st iceberg.StructType, v []func() bool) bool {
-	out := true
-	for _, res := range v {
-		out = res() && out
-	}
-
-	return out
-}
-
-func (sc *schemaCompatVisitor) Field(n iceberg.NestedField, v func() bool) bool {
-	return sc.isFieldCompat(n) && v()
-}
-
-func (sc *schemaCompatVisitor) List(l iceberg.ListType, v func() bool) bool {
-	return sc.isFieldCompat(l.ElementField()) && v()
-}
-
-func (sc *schemaCompatVisitor) Map(m iceberg.MapType, vk, vv func() bool) bool {
-	return sc.isFieldCompat(m.KeyField()) && sc.isFieldCompat(m.ValueField()) && vk() && vv()
-}
-
-func (sc *schemaCompatVisitor) Primitive(p iceberg.PrimitiveType) bool {
-	return true
-}
 
 type snapshotUpdate struct {
 	txn           *Transaction
@@ -174,10 +39,6 @@ type snapshotUpdate struct {
 func (s snapshotUpdate) fastAppend() *snapshotProducer {
 	return newFastAppendFilesProducer(OpAppend, s.txn, s.io, nil, s.snapshotProps)
 }
-
-// func (s snapshotUpdate) mergeAppend() *snapshotProducer {
-// 	return newMergeAppendFilesProducer(OpAppend, s.txn, s.io, nil, s.snapshotProps)
-// }
 
 type Transaction struct {
 	tbl  *Table
@@ -242,17 +103,6 @@ func (t *Transaction) updateSnapshot(props iceberg.Properties) snapshotUpdate {
 		snapshotProps: props,
 	}
 }
-
-// func (t *Transaction) appendSnapshotProducer(props iceberg.Properties) *snapshotProducer {
-// 	manifestMerge := t.meta.props.GetBool(ManifestMergeEnabledKey, ManifestMergeEnabledDefault)
-
-// 	updateSnapshot := t.updateSnapshot(props)
-// 	if manifestMerge {
-// 		return updateSnapshot.mergeAppend()
-// 	}
-
-// 	return updateSnapshot.fastAppend()
-// }
 
 func (t *Transaction) SetProperties(props iceberg.Properties) error {
 	if len(props) > 0 {
